backend/requirements.txt
fastapi
uvicorn[standard]
spacy
vaderSentiment
transformers
torch
openai
anthropic
python-dotenv
websockets
httpx
